+++ tests/testthat/test-decorated-storr.R
@@ -17,6 +17,14 @@ test_with_dir("empty keys with decorated storr", {
   expect_equal(cache$decode_namespaced(x), x)
 })
 
+test_with_dir("drake_tempfile()", {
+  expect_error(drake_tempfile())
+  cache <- new_cache()
+  x <- drake_tempfile(cache = cache)
+  expect_true(grepl("drake", x))
+  expect_true(grepl("tmp", x))
+})
+
 test_with_dir("key encoding for paths and namespaced functions", {
   x <- c("myfunny:::variablename", "relative/path\na\\m//e")
   expect_false(all(is_encoded_path(x)))
@@ -125,14 +133,12 @@ test_with_dir("run through non-encoder decorated storr methods", {
   x$exists_object("x")
   y <- storr::storr_environment()
   x$set("x", 1)
-  x$export(y, "x")
   x$fill("z", 2)
   x$flush_cache()
   x$get("x")
   x$get_value(x$get_hash("x"))
   x$hash_object("x")
   x$hash_raw("x")
-  x$import(y)
   x$index_export()
   x$index_import(x$index_export()[1:2, ])
   x$list()
@@ -417,6 +423,50 @@ test_with_dir("fst_dt format forces data.tables", {
   expect_true(inherits(readd(x), "data.table"))
 })
 
+test_with_dir("disk.frame (#1004)", {
+  skip_if_not_installed("disk.frame")
+  skip_if_not_installed("fst")
+  plan <- drake_plan(
+    x = target(
+      disk.frame::as.disk.frame(
+        data.frame(x = letters, y = letters, stringsAsFactors = FALSE),
+        outdir = drake_tempfile()
+      ),
+      format = "diskframe"
+    )
+  )
+  make(plan)
+  out <- readd(x)
+  expect_true(inherits(out, "disk.frame"))
+  exp <- data.table::as.data.table(
+    data.frame(x = letters, y = letters, stringsAsFactors = FALSE)
+  )
+  expect_equivalent(as.data.frame(out), exp)
+  cache <- drake_cache()
+  expect_equivalent(
+    as.data.frame(cache$get_value(cache$get_hash("x"))),
+    exp
+  )
+  ref <- cache$storr$get("x")
+  expect_true(inherits(ref, "drake_format_diskframe"))
+  expect_equal(length(ref), 1L)
+  expect_true(nchar(ref) < 100)
+  expect_false(is.list(ref))
+})
+
+test_with_dir("diskframe format forces disk.frames", {
+  skip_if_not_installed("disk.frame")
+  skip_if_not_installed("fst")
+  plan <- drake_plan(
+    x = target(
+      data.frame(x = letters, y = letters),
+      format = "diskframe"
+    )
+  )
+  expect_warning(make(plan), regexp = "disk.frame")
+  expect_true(inherits(readd(x), "disk.frame"))
+})
+
 test_with_dir("drop format for NULL values (#998)", {
   skip_if(getRversion() < "3.5.0")
   f <- function() {
@@ -426,4 +476,104 @@ test_with_dir("drop format for NULL values (#998)", {
   make(plan)
   expect_null(readd(x))
   expect_null(drake_cache()$storr$get("x"))
+})
+
+test_with_dir("decorated storr import (#1015)", {
+  skip_if(getRversion() < "3.5.0")
+  plan1 <- drake_plan(
+    w = "w",
+    x = "x",
+    y = target("y", format = "rds"),
+    z = target("z", format = "rds")
+  )
+  plan2 <- drake_plan(
+    a = "a",
+    x = "x",
+    b = target("b", format = "rds"),
+    z = target("z2", format = "rds")
+  )
+  cache1 <- new_cache("cache1")
+  cache2 <- new_cache("cache2")
+  make(plan1, cache = cache1)
+  make(plan2, cache = cache2)
+  cache1$import(cache2)
+  expect_equal(cache1$get("a"), "a")
+  expect_true(is.list(cache1$get("a", namespace = "meta")))
+  expect_equal(cache1$get("b"), "b")
+  expect_equal(cache1$get("z"), "z2")
+})
+
+test_with_dir("decorated storr export (#1015)", {
+  skip_if(getRversion() < "3.5.0")
+  plan1 <- drake_plan(
+    w = "w",
+    x = "x",
+    y = target("y", format = "rds"),
+    z = target("z", format = "rds")
+  )
+  plan2 <- drake_plan(
+    a = "a",
+    x = "x",
+    b = target("b", format = "rds"),
+    z = target("z2", format = "rds")
+  )
+  cache1 <- new_cache("cache1")
+  cache2 <- new_cache("cache2")
+  make(plan1, cache = cache1)
+  make(plan2, cache = cache2)
+  cache2$export(cache1, gc = FALSE)
+  expect_equal(cache1$get("a"), "a")
+  expect_true(is.list(cache1$get("a", namespace = "meta")))
+  expect_equal(cache1$get("b"), "b")
+  expect_equal(cache1$get("z"), "z2")
+})
+
+test_with_dir("decorated storr import specific targets (#1015)", {
+  skip_if(getRversion() < "3.5.0")
+  plan1 <- drake_plan(
+    w = "w",
+    x = "x",
+    y = target("y", format = "rds"),
+    z = target("z", format = "rds")
+  )
+  plan2 <- drake_plan(
+    a = "a",
+    x = "x",
+    b = target("b", format = "rds"),
+    z = target("z2", format = "rds")
+  )
+  cache1 <- new_cache("cache1")
+  cache2 <- new_cache("cache2")
+  make(plan1, cache = cache1)
+  make(plan2, cache = cache2)
+  cache1$import(cache2, a)
+  expect_equal(cache1$get("a"), "a")
+  expect_true(is.list(cache1$get("a", namespace = "meta")))
+  expect_false(cache1$exists("b"))
+  expect_equal(cache1$get("z"), "z")
+})
+
+test_with_dir("decorated storr import specific targets (#1015)", {
+  skip_if(getRversion() < "3.5.0")
+  plan1 <- drake_plan(
+    w = "w",
+    x = "x",
+    y = target("y", format = "rds"),
+    z = target("z", format = "rds")
+  )
+  plan2 <- drake_plan(
+    a = "a",
+    x = "x",
+    b = target("b", format = "rds"),
+    z = target("z2", format = "rds")
+  )
+  cache1 <- new_cache("cache1")
+  cache2 <- new_cache("cache2")
+  make(plan1, cache = cache1)
+  make(plan2, cache = cache2)
+  cache2$export(cache1, a)
+  expect_equal(cache1$get("a"), "a")
+  expect_true(is.list(cache1$get("a", namespace = "meta")))
+  expect_false(cache1$exists("b"))
+  expect_equal(cache1$get("z"), "z")
 })